refactor: reducer 파일 분리
- cartProdcuts, checkedIds, product, products, snackBar 파일로 분리
