+++ src/index.js
@@ -4,11 +4,11 @@ import App from "./App";
 import { BrowserRouter } from "react-router-dom";
 import GlobalStyles from "./globalStyles";
 import { applyMiddleware, createStore } from "redux";
-import appReducer from "./modules/products";
 import { composeWithDevTools } from "redux-devtools-extension";
 import { Provider } from "react-redux";
 import createThunkMiddleware from "./lib/thunk";
 import ThemeProvider from "./ThemeProvider";
+import appReducer from "./modules";
 
 const store = createStore(
   appReducer,
