+++ src/modules/checkedIds.js
@@ -1,0 +1,56 @@
+import createReducer from "./createReducer";
+
+const ADD_PRODUCT_ID = "checked-ids/ADD_PRODUCT_ID";
+export const REMOVE_PRODUCT_ID = "checked-ids/REMOVE_PRODUCT_ID";
+export const ADD_PRODUCT_IDS = "checked-ids/ADD_PRODUCT_IDS";
+const REMOVE_PRODUCT_IDS = "checked-ids/REMOVE_PRODUCT_IDS";
+
+export const removeIds = () => ({
+  type: REMOVE_PRODUCT_IDS,
+});
+
+export const addIds = (ids) => ({
+  type: ADD_PRODUCT_IDS,
+  ids,
+});
+
+export const removeId = (id) => ({
+  type: REMOVE_PRODUCT_ID,
+  removeId: id,
+});
+
+export const addId = (id) => ({
+  type: ADD_PRODUCT_ID,
+  newId: id,
+});
+
+const removeProductId = (state, action) => {
+  const removeIndex = state.findIndex((id) => id === action.removeId);
+  const newState = [...state];
+  newState.splice(removeIndex, 1);
+  return newState;
+};
+
+const addProductId = (state, action) => {
+  const addIndex = state.findIndex((id) => id === action.newId);
+  if (addIndex === -1) {
+    return [...state, action.newId];
+  }
+  return [...state];
+};
+
+const removeProductIds = () => [];
+
+const addProductIds = (_, action) => action.ids;
+
+const checkedProductIdsReducer = createReducer(
+  {},
+  {
+    [ADD_PRODUCT_ID]: addProductId,
+    [REMOVE_PRODUCT_ID]: removeProductId,
+    [ADD_PRODUCT_IDS]: addProductIds,
+    [REMOVE_PRODUCT_IDS]: removeProductIds,
+  }
+);
+
+export default checkedProductIdsReducer;
